proc_cpulim: retain and release dispatch_queue(s) on-app-level (before sending release to libdispatch). Fixing 'signed' warnings.

+++ ChromeMenu/src/CMScrollView.m
@@ -29,7 +29,7 @@ const short down = 1;
 	CGFloat amount;
 	short direction;
 	
-	multiplier = ceil(ABS(deltaY) / 2);
+	multiplier = (int)ceil(ABS(deltaY) / 2);
 	amount = multiplier * lineHeight;
 	direction = (deltaY < 0) ? down : up;
